Win32/MSVC2015 build and warnings fixes

+++ src/actors/robot/robotmodule.h
@@ -507,7 +507,7 @@ namespace ActorRobot {
         QToolButton * textEditBtn;
         QToolButton * radEditBtn;
         QToolButton * tmpEditBtn;
-        float c_scale;
+        double c_scale;
         int CurCellSize;
         
     };
