+++ src/plugins/kumircoderun/runplugin.cpp
@@ -120,7 +120,7 @@ QMap<QString,QVariant> KumirRunPlugin::getScalarLocalValues(int frameNo) const
     pRun_->lockVMMutex();
     QMap<QString,QVariant> result;
     const std::vector<Variable> & locals = pRun_->vm->getLocals(frameNo);
-    for (int i=0; i<locals.size(); i++) {
+    for (int i=0; i<(int)locals.size(); i++) {
         const Variable & var = locals.at(i);
         if (var.dimension()==0) {
             const QString varName = QString::fromStdWString(var.myName());
@@ -148,7 +148,7 @@ QMap<QString,QVariant> KumirRunPlugin::getScalarGlobalValues(const QString & mod
     pRun_->lockVMMutex();
     QMap<QString,QVariant> result;
     const std::vector<Variable> & locals = pRun_->vm->getGlobals(moduleName.toStdWString());
-    for (int i=0; i<locals.size(); i++) {
+    for (int i=0; i<(int)locals.size(); i++) {
         const Variable & var = locals.at(i);
         if (var.dimension()==0) {
             const QString varName = QString::fromStdWString(var.myName());
@@ -286,7 +286,7 @@ QVariantList KumirRunPlugin::getLocalTableValues(
     int counter = 0;
     pRun_->lockVMMutex();
     const std::vector<Variable> & locals = pRun_->vm->getLocals(frameNo);
-    for (int i=0; i<locals.size(); i++) {
+    for (int i=0; i<(int)locals.size(); i++) {
         const Variable & var = locals.at(i);
         if (var.dimension()>0 && var.myName()==name.toStdWString()) {
             complete = true;
@@ -312,7 +312,7 @@ QVariant KumirRunPlugin::getLocalTableValue(
     }
     pRun_->lockVMMutex();
     const std::vector<Variable> & locals = pRun_->vm->getLocals(frameNo);
-    for (int i=0; i<locals.size(); i++) {
+    for (int i=0; i<(int)locals.size(); i++) {
         const Variable & var = locals.at(i);
         if (var.dimension()>0 && var.myName()==name.toStdWString()) {
             if (var.hasValue(kIndeces))
@@ -336,7 +336,7 @@ QVariantList KumirRunPlugin::getGlobalTableValues(
     int counter = 0;
     pRun_->lockVMMutex();
     const std::vector<Variable> & globals = pRun_->vm->getGlobals(moduleName.toStdWString());
-    for (int i=0; i<globals.size(); i++) {
+    for (int i=0; i<(int)globals.size(); i++) {
         const Variable & var = globals.at(i);
         if (var.dimension()>0 && var.myName()==name.toStdWString()) {
             complete = true;
@@ -362,7 +362,7 @@ QVariant KumirRunPlugin::getGlobalTableValue(
     }
     pRun_->lockVMMutex();
     const std::vector<Variable> & globals = pRun_->vm->getGlobals(moduleName.toStdWString());
-    for (int i=0; i<globals.size(); i++) {
+    for (int i=0; i<(int)globals.size(); i++) {
         const Variable & var = globals.at(i);
         if (var.dimension()>0 && var.myName()==name.toStdWString()) {
             if (var.hasValue(kIndeces))
